+++ src/client/cypress/e2e/detailPage/location.cy.js
@@ -1,6 +1,6 @@
 import { addItem, saveWithSaveBar, stopEditing } from "../helpers/buttonHelpers";
 import { checkRowWithText, clickOnRowWithText, showTableAndWaitForData } from "../helpers/dataGridHelpers";
-import { evaluateInput, evaluateSelect, isDisabled, setInput, setSelect } from "../helpers/formHelpers";
+import { clearInput, evaluateInput, evaluateSelect, isDisabled, setInput, setSelect } from "../helpers/formHelpers";
 import {
   createBorehole,
   goToRouteAndAcceptTerms,
@@ -66,31 +66,59 @@ describe("Tests for 'Location' edit page.", () => {
     );
     cy.get("@borehole_id").then(id => {
       goToRouteAndAcceptTerms(`/${id}`);
-      cy.get('[data-cy="originalName-formInput"]').within(() => {
-        cy.get("input").as("originalNameInput");
-      });
-      cy.get('[data-cy="name-formInput"]').within(() => {
-        cy.get("input").as("nameInput");
-      });
 
-      cy.get("@originalNameInput").should("have.value", "PHOTOSQUIRREL");
-      cy.get("@nameInput").should("have.value", "PHOTOSQUIRREL");
+      evaluateInput("originalName", "PHOTOSQUIRREL");
+      evaluateInput("name", "PHOTOSQUIRREL");
 
       startBoreholeEditing();
       // changing original name should also change alternate name
-      cy.get("@originalNameInput").clear().type("PHOTOCAT");
-      cy.get("@originalNameInput").should("have.value", "PHOTOCAT");
-      cy.get("@nameInput").should("have.value", "PHOTOCAT");
+      setInput("originalName", "PHOTOCAT");
+      evaluateInput("originalName", "PHOTOCAT");
+      evaluateInput("name", "PHOTOCAT");
 
-      cy.get("@nameInput").clear().type("PHOTOMOUSE");
-      cy.get("@originalNameInput").should("have.value", "PHOTOCAT");
-      cy.get("@nameInput").should("have.value", "PHOTOMOUSE");
+      // changing alternate name should not change original name
+      setInput("name", "PHOTOMOUSE");
+      evaluateInput("originalName", "PHOTOCAT");
+      evaluateInput("name", "PHOTOMOUSE");
 
-      cy.get("@nameInput").clear();
+      // changing original name should not update alternate name if they are different
+      setInput("originalName", "PHOTOPIGEON");
+      evaluateInput("originalName", "PHOTOPIGEON");
+      evaluateInput("name", "PHOTOMOUSE");
+
+      clearInput("name");
+      evaluateInput("originalName", "PHOTOPIGEON");
+      evaluateInput("name", "");
       saveWithSaveBar();
       // should be reset to original name if alternate name is empty
-      cy.get("@originalNameInput").should("have.value", "PHOTOCAT");
-      cy.get("@nameInput").should("have.value", "PHOTOCAT");
+      evaluateInput("originalName", "PHOTOPIGEON");
+      evaluateInput("name", "PHOTOPIGEON");
+
+      // should keep different alternate name when switching tabs
+      setInput("name", "PHOTOMOUSE");
+      saveWithSaveBar();
+      cy.get('[data-cy="borehole-menu-item"]').click();
+      cy.get('[data-cy="location-menu-item"]').click();
+      evaluateInput("originalName", "PHOTOPIGEON");
+      evaluateInput("name", "PHOTOMOUSE");
+    });
+  });
+
+  it("does not overwrite alternate name if it is different from original name", () => {
+    createBorehole({ "extended.original_name": "PHOTOSQUIRREL", "custom.alternate_name": "PHOTOMOUSE" }).as(
+      "borehole_id",
+    );
+    cy.get("@borehole_id").then(id => {
+      goToRouteAndAcceptTerms(`/${id}`);
+
+      evaluateInput("originalName", "PHOTOSQUIRREL");
+      evaluateInput("name", "PHOTOMOUSE");
+
+      startBoreholeEditing();
+      // changing original name should not update alternate name if they are different
+      setInput("originalName", "PHOTOPIGEON");
+      evaluateInput("originalName", "PHOTOPIGEON");
+      evaluateInput("name", "PHOTOMOUSE");
     });
   });