+++ tests/api/Controllers/ImportControllerTest.cs
@@ -108,10 +108,7 @@ public class ImportControllerTest
         OkObjectResult okResult = (OkObjectResult)response.Result!;
         Assert.AreEqual(2, okResult.Value);
 
-        var boreholes = await GetBoreholesWithIncludes(context.Boreholes).ToListAsync().ConfigureAwait(false);
-
-        var borehole = boreholes.Find(b => b.OriginalName == "PURPLETOLL");
-
+        var borehole = await context.Boreholes.GetAllWithIncludes().SingleAsync(b => b.OriginalName == "PURPLETOLL").ConfigureAwait(false);
         Assert.IsNotNull(borehole.CreatedById, nameof(Borehole.CreatedById).ShouldNotBeNullMessage());
         Assert.IsNotNull(borehole.Created, nameof(Borehole.Created).ShouldNotBeNullMessage());
         Assert.IsNotNull(borehole.UpdatedById, nameof(Borehole.UpdatedById).ShouldNotBeNullMessage());
@@ -554,7 +551,7 @@ public class ImportControllerTest
         ActionResultAssert.IsOk(response.Result);
         OkObjectResult okResult = (OkObjectResult)response.Result!;
         Assert.AreEqual(2, okResult.Value);
-        var uploadedBoreholesWithAttachment = await GetBoreholesWithIncludes(context.Boreholes).Where(b => b.OriginalName.StartsWith("Carmen Catnip")).ToListAsync();
+        var uploadedBoreholesWithAttachment = await context.Boreholes.GetAllWithIncludes().Where(b => b.OriginalName.StartsWith("Carmen Catnip")).ToListAsync();
         Assert.AreEqual(uploadedBoreholesWithAttachment.SelectMany(b => b.Files!).Count(), 3);
 
         var firstBoreholes = uploadedBoreholesWithAttachment.Find(b => b.OriginalName == "Carmen Catnip Cheese");
@@ -638,7 +635,7 @@ public class ImportControllerTest
         Assert.AreEqual(6, okResult.Value);
 
         // Assert imported values
-        var borehole = GetBoreholesWithIncludes(context.Boreholes).ToList().Find(b => b.OriginalName == "Unit_Test_6");
+        var borehole = await context.Boreholes.GetAllWithIncludes().SingleAsync(b => b.OriginalName == "Unit_Test_6");
         Assert.AreEqual(1, borehole.WorkgroupId);
         Assert.AreEqual("Unit_Test_6_a", borehole.Name);
         Assert.AreEqual(null, borehole.IsPublic);
@@ -686,7 +683,7 @@ public class ImportControllerTest
         Assert.AreEqual(6, okResult.Value);
 
         // Assert imported values
-        var borehole = GetBoreholesWithIncludes(context.Boreholes).ToList().Find(b => b.OriginalName == "Unit_Test_2");
+        var borehole = await context.Boreholes.GetAllWithIncludes().SingleAsync(b => b.OriginalName == "Unit_Test_2");
         Assert.AreEqual(1, borehole.WorkgroupId);
         Assert.AreEqual(null, borehole.Name);
         Assert.AreEqual(null, borehole.IsPublic);
@@ -725,7 +722,7 @@ public class ImportControllerTest
         Assert.AreEqual(7, okResult.Value);
 
         // Assert imported values
-        var boreholeLV95 = GetBoreholesWithIncludes(context.Boreholes).ToList().Find(b => b.OriginalName == "Unit_Test_2");
+        var boreholeLV95 = await context.Boreholes.GetAllWithIncludes().SingleAsync(b => b.OriginalName == "Unit_Test_2");
         Assert.AreEqual(ReferenceSystem.LV95, boreholeLV95.OriginalReferenceSystem);
         Assert.AreEqual(2000010.12, boreholeLV95.LocationX);
         Assert.AreEqual(1000010.1, boreholeLV95.LocationY);
@@ -734,7 +731,7 @@ public class ImportControllerTest
         Assert.AreEqual(2, boreholeLV95.PrecisionLocationXLV03);
         Assert.AreEqual(2, boreholeLV95.PrecisionLocationYLV03);
 
-        var boreholeLV03 = GetBoreholesWithIncludes(context.Boreholes).ToList().Find(b => b.OriginalName == "Unit_Test_6");
+        var boreholeLV03 = await context.Boreholes.GetAllWithIncludes().SingleAsync(b => b.OriginalName == "Unit_Test_6");
         Assert.AreEqual(ReferenceSystem.LV03, boreholeLV03.OriginalReferenceSystem.Value);
         Assert.AreEqual(20050.12, boreholeLV03.LocationXLV03);
         Assert.AreEqual(10050.12345, boreholeLV03.LocationYLV03);
@@ -743,7 +740,7 @@ public class ImportControllerTest
         Assert.AreEqual(5, boreholeLV03.PrecisionLocationX);
         Assert.AreEqual(5, boreholeLV03.PrecisionLocationY);
 
-        var boreholeWithZeros = GetBoreholesWithIncludes(context.Boreholes).ToList().Find(b => b.OriginalName == "Unit_Test_7");
+        var boreholeWithZeros = await context.Boreholes.GetAllWithIncludes().SingleAsync(b => b.OriginalName == "Unit_Test_7");
         Assert.AreEqual(ReferenceSystem.LV03, boreholeWithZeros.OriginalReferenceSystem.Value);
         Assert.AreEqual(20060.000, boreholeWithZeros.LocationXLV03);
         Assert.AreEqual(10060.0000, boreholeWithZeros.LocationYLV03);
@@ -1089,18 +1086,6 @@ public class ImportControllerTest
 
         ValidationProblemDetails problemDetails = GetProblemDetailsFromResponse(response);
         Assert.AreEqual(1000, problemDetails.Errors.Count);
-    }
-
-    [TestMethod]
-    public void CompareValueWithTolerance()
-    {
-        Assert.AreEqual(true, ImportController.CompareValuesWithTolerance(null, null, 0));
-        Assert.AreEqual(true, ImportController.CompareValuesWithTolerance(2100000, 2099998, 2));
-        Assert.AreEqual(false, ImportController.CompareValuesWithTolerance(2100000, 2000098, 1.99));
-        Assert.AreEqual(false, ImportController.CompareValuesWithTolerance(2100002, 2000000, 1.99));
-        Assert.AreEqual(false, ImportController.CompareValuesWithTolerance(21000020, 2000000, 20));
-        Assert.AreEqual(false, ImportController.CompareValuesWithTolerance(null, 2000000, 0));
-        Assert.AreEqual(false, ImportController.CompareValuesWithTolerance(2000000, null, 2));
     }
 
     [TestMethod]