+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 ## [Unreleased]
 
+### Changed
+- Attachments are now displayed in a table following new UI design.
+
+### Fixed
+- Attachments page was missing scrollbars.
+- The input field for the alternate name was always overwritten by the original name when loading the location page.
+- Tooltips in the labeling area where not complete and would sometimes overlap with text to be selected.
+- Copy borehole to the correct workgroup if the user only has edit permissions in a single workgroup.
+
+## v2.1.1116 - 2025-03-17
+
+### Changed
+
+- Increased Disclaimer dialog width.
+
+## v2.1.1113 - 2025-03-17
+
 ### Added
 
 - Added a table displaying all the users in the admin settings.
@@ -10,6 +27,7 @@
 - Added a detail view for workgroups in the admin settings.
 - Enable auto setting of depth in MASL in hydro module.
 - Added functionality to copy extracted text to clipboard in the labeling panel.
+- New _extern-sync_ Docker image for syncing published boreholes from a source to a target database.
 - Added highlighting of words that will be included in text extraction.
 - Added data collection with consent option.
 
@@ -28,6 +46,8 @@
 - When hovering over the main side navigation icons, an incorrect background color was displayed for the hovered icon.
 - Before export or import check if the user has the correct role and is in the same workgroup as the borehole.
 - Fix sorting by drilling purpose in borehole table.
+- Fix unexpected reset of column width in borehole table.
+- Stop automatic update of alternate name when it is different from the original name.
 
 ## v2.1.1052 - 2025-02-04